Add documentation for extension tests (#5)
* Move slicing functions from `extension.test.ts` to `slice_helpers.ts`

* Add documentation for slice helper functions

* Add documentation for `proxy.ts`

* Add comments to `runTests.ts`
